Pull JSON parsing out of AddonInfo
This is driven by a new requirement (multiple versions in one remote .version file).

It should be the case that all the error paths are pretty much the same.  Only major difference is that if the parsed JSON does not return a dict there will be a handled cast exception rather than a generic message that the file has errors.